Properties New component added as a Properties advanced stories
