--- raaghu-components/src/rds-comp-properties-new/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./rds-comp-properties-new";
