--- raaghu-components/src/rds-comp-properties-new/rds-comp-properties-new.module.css
@@ -1,1 +1,0 @@
-.RdsCompPropertiesNew {}
