--- raaghu-components/src/rds-comp-properties-new/rds-comp-properties-new.stories.tsx
@@ -1,27 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import RdsCompPropertiesNew from './rds-comp-properties-new';
-
-
-const meta: Meta = { 
-    title: "Components/Properties New",
-    component: RdsCompPropertiesNew,
-    parameters: {
-        layout: 'padded',
-    },
-    tags: ['autodocs'],
-    argTypes: {
-    },
-} satisfies Meta<typeof RdsCompPropertiesNew>;
-
-export default meta;
-type Story = StoryObj<typeof RdsCompPropertiesNew>;
-
-export const Default: Story = {
-    args: {
-        
-    }
-} satisfies Story;
-
-
-
-
