+++ raaghu-components/src/rds-comp-properties/rds-comp-properties.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import RdsCompDatatable from "../rds-comp-data-table/rds-comp-data-table";
-import { RdsInput, RdsButton } from "../rds-elements";
+import { RdsInput, RdsButton, RdsIcon } from "../rds-elements";
 import { useTranslation } from "react-i18next";
 
 export interface RdsCompPropertiesProp {
@@ -17,12 +17,18 @@ export interface RdsCompPropertiesProp {
         isEndUserEditing?: boolean;
     }[];
     onActionSelection(arg: any): void;
+    displayType: "basic" | "advanced";
 }
 
 const RdsCompProperties = (props: RdsCompPropertiesProp) => {
 
     const [key, setKey] = useState("");
     const [value, setValue] = useState("");
+    const [tableData, setTableData] = useState<any>([]);
+    const [propertyData, setPropertyData] = useState<any>({
+        key: "",
+        PropValue: "",
+    });
 
     const isHeaderKeyValid = (headerKey: any) => {
         if (!headerKey || headerKey.length === 0) {
@@ -48,6 +54,13 @@ const RdsCompProperties = (props: RdsCompPropertiesProp) => {
         setValue(event.target.value);
     };
 
+    function headerKeyhandleChanged(value: any) {
+        setPropertyData({ ...propertyData, key: value });
+    }
+    function headerValuehandleChanged(value: any) {
+        setPropertyData({ ...propertyData, PropValue: value });
+    }
+
     //****************handle Submit********************
     const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
@@ -55,8 +68,64 @@ const RdsCompProperties = (props: RdsCompPropertiesProp) => {
         setKey("");
         setValue("");
     };
+
+
+    const tableHeaders = [
+        {
+            displayName: "Key",
+            key: "key",
+            datatype: "text",
+            sortable: true,
+        },
+        {
+            displayName: "Value",
+            key: "Value",
+            datatype: "text",
+            sortable: true,
+        },
+        {
+            displayName: "Action",
+            key: "delete",
+            datatype: "children",
+            sortable: true,
+        },
+    ];
+    const handleAddItem = () => {
+        let newTempData: any;
+        newTempData = {
+            key: propertyData.key,
+            Value: propertyData.PropValue,
+            delete: (
+                <>
+                    <a
+                        className="mx-2"
+                        type="button"
+                        data-bs-toggle="modal"
+                        data-bs-target={"#deleteTreeNode"}
+                        onClick={() => onDelete(propertyData.key)}
+                        data-testid="delete"
+                    >
+                        <RdsIcon
+                            name={"delete"}
+                            height="16px"
+                            width="20px"
+                            stroke={true}
+                        ></RdsIcon>
+                    </a>
+                </>
+            ),
+        };
+        setTableData((prev: any) => [...prev, newTempData]);
+        setPropertyData({ key: "", PropValue: "" });
+    };
+    function onDelete(key: any) {
+        const tempPropertyData = tableData.filter((el: any) => el.key != key);
+        setTableData(tempPropertyData);
+    }
+    
     return (
         <>
+        {props.displayType == "basic" && (
             <form onSubmit={handleSubmit}>
                 <RdsCompDatatable
                     actionPosition="right"
@@ -121,6 +190,77 @@ const RdsCompProperties = (props: RdsCompPropertiesProp) => {
                         ></RdsButton>
                 </div>
             </form>
+        )}
+        {props.displayType === "advanced" && (
+            <form onSubmit={handleSubmit}>                
+            <div className=" row g-3 mt-3 px-2">
+                <div className="col-md-5">
+                    <RdsInput
+                        label="Key"
+                        placeholder="Enter a key"
+                        inputType="text"
+                        onChange={(e: any) => {
+                            headerKeyhandleChanged(e.target.value);
+                        }}
+                        value={propertyData.key}
+                        dataTestId="key"
+                    ></RdsInput>
+                </div>
+                <div className="col-md-5">
+                    <RdsInput
+                        label="Value"
+                        placeholder="Enter a value"
+                        inputType="text"
+                        onChange={(e: any) => {
+                            headerValuehandleChanged(e.target.value);
+                        }}
+                        value={propertyData.PropValue}
+                        dataTestId="value"
+                    ></RdsInput>
+                </div>
+                <div className="col-md-2   align-items-end d-flex">                    
+                    <RdsButton 
+                        label="Add"                                              
+                        colorVariant="primary"
+                       block={false}
+                       onClick={handleAddItem}
+                        tooltipTitle={""}
+                        type="submit"
+                        dataTestId="add"
+                    />
+                </div>
+            </div>
+            <div className="mt-3 px-2">
+                <RdsCompDatatable
+                    actionPosition="right"
+                    tableHeaders={tableHeaders}
+                    tableData={tableData}
+                    pagination={false}
+                ></RdsCompDatatable>
+            </div>
+            <div className=" mt-5 pb-3 p-4 footer-buttons d-flex gap-2 flex-column-reverse flex-lg-row flex-md-column-reverse flex-xl-row flex-xxl-row flex-row">
+                <RdsButton
+                    label="Cancel"
+                    colorVariant="primary"
+                    block={false}
+                    tooltipTitle={""}
+                    size="small"
+                    type="button"
+                    isOutline={true}
+                    dataTestId="cancel"
+                />
+                <RdsButton
+                    label="Save"
+                    colorVariant="primary"
+                    block={false}
+                    tooltipTitle={""}
+                    size="small"
+                    type="submit"
+                    dataTestId="submit"
+                />
+            </div>
+        </form>
+        )}
         </>
     );
 };
